% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fies.R
\name{calc_fies}
\alias{calc_fies}
\title{Calculer le score FIES (Food Insecurity Experience Scale)}
\usage{
calc_fies(data)
}
\arguments{
\item{data}{Le dataframe issu de la base S08 du Niger.}
}
\value{
Un dataframe avec les colonnes \code{grappe}, \code{menage}, \code{fies_score},
\code{fies_moderee} et \code{fies_severe}.
}
\description{
Calculer le score FIES (Food Insecurity Experience Scale)
}
