% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hdds.R
\name{calc_hdds}
\alias{calc_hdds}
\title{Calculer le score de diversité alimentaire (HDDS)}
\usage{
calc_hdds(data)
}
\arguments{
\item{data}{Le dataframe issu de la base S7B (ex: b4).}
}
\value{
Un dataframe avec les colonnes \code{grappe}, \code{menage} et \code{score_hdds}.
}
\description{
Calculer le score de diversité alimentaire (HDDS)
}
