% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_filiere.R
\name{load_filiere}
\alias{load_filiere}
\title{Charger les bases de données de l'EHCVM Niger}
\usage{
load_filiere(data_dir = here("data"))
}
\arguments{
\item{data_dir}{Le chemin du dossier contenant les fichiers \code{.dta}.
Par défaut, \code{here("data")}.}
}
\value{
Une liste nommée contenant les 9 bases de données (b1 à b9).
}
\description{
Charger les bases de données de l'EHCVM Niger
}
