% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prix_chaine.R
\name{prix_chaine}
\alias{prix_chaine}
\title{Analyse de la chaîne des prix et des marges commerciales}
\usage{
prix_chaine(data_ventes, data_prix_marche, data_qc_s2, data_qc_s3)
}
\arguments{
\item{data_ventes}{Base des ventes (b7 - s16d_me_ner2021.dta).}

\item{data_prix_marche}{Base des prix communautaires (QC-S5).}

\item{data_qc_s2}{Base communautaire S2 (distances).}

\item{data_qc_s3}{Base communautaire S3 (coopératives).}
}
\value{
Une liste contenant \code{marge_region} (tableau) et \code{model} (objet de régression).
}
\description{
Calcule le prix producteur, le prix du marché communautaire, la marge
commerciale, et effectue une régression des déterminants du prix producteur
(distance au marché, présence de coopérative).
}
