% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profil_menage.R
\name{profil_menage}
\alias{profil_menage}
\title{Profilage des ménages (Typologie 4 groupes et indicateurs)}
\usage{
profil_menage(
  data_conso,
  data_prod,
  data_welfare,
  data_demo,
  data_fies,
  code_prod = 7,
  code_cult = 1
)
}
\arguments{
\item{data_conso}{La base de consommation (b4 - s07b_me_ner2021.dta).}

\item{data_prod}{La base de production (b6 - s16c_me_ner2021.dta).}

\item{data_welfare}{La base de pondération et milieu (b2 - ehcvm_conso_ner2021.dta).}

\item{data_demo}{La base démographique (s01_me_ner2021.dta).}

\item{data_fies}{La base FIES (s08a_me_ner2021.dta).}

\item{code_prod}{Le code produit du Mil dans la base consommation (défaut = 7).}

\item{code_cult}{Le code culture du Mil dans la base production (défaut = 1).}
}
\value{
Un dataframe contenant les statistiques agrégées par groupe typologique.
}
\description{
Cette fonction construit la typologie Producteur/Consommateur pour le produit
stratégique, fusionne les données démographiques (S01), de consommation (S7B),
de production (S16C) et de sécurité alimentaire (FIES) pour créer un tableau
de profilage complet.
}
