% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_filiere.R
\name{reg_filiere}
\alias{reg_filiere}
\title{Régression d'impact de la filière sur la sécurité alimentaire}
\usage{
reg_filiere(data_fies, data_hdds, data_ventes, data_prod, data_welfare)
}
\arguments{
\item{data_fies}{Base FIES.}

\item{data_hdds}{Base HDDS (produit par calc_hdds()).}

\item{data_ventes}{Base S16D (ventes et revenus).}

\item{data_prod}{Base S16C (producteur).}

\item{data_welfare}{Base des contrôles (taille, région, éducation, etc.).}
}
\value{
Une liste contenant les modèles \code{ols}, \code{iv}, \code{hdds_model}.
}
\description{
Estime l'impact du statut de producteur de Mil sur le score FIES en utilisant
une méthode des variables instrumentales (2SLS) pour corriger l'endogénéité,
ainsi que des modèles de robustesse (HDDS) et des interactions spatiales.
}
